+++ python/metatensor_core/metatensor/io/__init__.py
@@ -22,6 +22,7 @@ from ._labels import (
     load_labels_buffer,
 )
 from ._mmap import load_block_mmap, load_mmap
+from ._partial import load_block_partial, load_partial
 from ._tensor import (
     _save_tensor,
     _save_tensor_buffer_raw,
@@ -40,12 +41,14 @@ __all__ = [
     "load_block_custom_array",
     "load_block",
     "load_block_mmap",
+    "load_block_partial",
     "load_buffer_custom_array",
     "load_buffer",
     "load_custom_array",
     "load_labels_buffer",
     "load_labels",
     "load_mmap",
+    "load_partial",
     "load",
 ]